# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API demo with auth",
        "title": "go-demo API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/sql-logs/report": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Aggregated anomalies and metrics within a time range. Defaults: last 7 days. Thresholds: slow_ms \u003e= 1000 OR (exec_time_ms \u003e= 500 AND exec_count \u003e= 100).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sql-logs"
                ],
                "summary": "SQL log report (JSON)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start time (RFC3339 or YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339 or YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by database name",
                        "name": "db",
                        "in": "query"
                    },
                    {
                        "maximum": 5000,
                        "minimum": 1,
                        "type": "integer",
                        "default": 500,
                        "description": "Max anomalies to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Slow threshold in ms",
                        "name": "slow_ms",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Frequent+slow time threshold in ms",
                        "name": "freq_slow_ms",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Frequent count threshold",
                        "name": "freq_count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Hard cap upper bound for anomalies count",
                        "name": "cap",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sqllog.ReportData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/sql-logs/report.csv": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Download the aggregated report as CSV.",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "sql-logs"
                ],
                "summary": "SQL log report (CSV)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start time (RFC3339 or YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339 or YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by database name",
                        "name": "db",
                        "in": "query"
                    },
                    {
                        "maximum": 5000,
                        "minimum": 1,
                        "type": "integer",
                        "default": 500,
                        "description": "Max anomalies to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Slow threshold in ms",
                        "name": "slow_ms",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Frequent+slow time threshold in ms",
                        "name": "freq_slow_ms",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Frequent count threshold",
                        "name": "freq_count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Hard cap upper bound for anomalies count",
                        "name": "cap",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/sql-logs/report.pdf": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Download the aggregated report as PDF.",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "sql-logs"
                ],
                "summary": "SQL log report (PDF)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start time (RFC3339 or YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339 or YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by database name",
                        "name": "db",
                        "in": "query"
                    },
                    {
                        "maximum": 5000,
                        "minimum": 1,
                        "type": "integer",
                        "default": 500,
                        "description": "Max anomalies to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Slow threshold in ms",
                        "name": "slow_ms",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Frequent+slow time threshold in ms",
                        "name": "freq_slow_ms",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Frequent count threshold",
                        "name": "freq_count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Hard cap upper bound for anomalies count",
                        "name": "cap",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "sqllog.AnomalyDetail": {
            "type": "object",
            "properties": {
                "db_name": {
                    "type": "string"
                },
                "exec_count": {
                    "type": "integer"
                },
                "exec_time_ms": {
                    "type": "integer"
                },
                "reasons": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sql_query": {
                    "type": "string"
                },
                "suggestions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "sqllog.ReportData": {
            "type": "object",
            "properties": {
                "anomalies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sqllog.AnomalyDetail"
                    }
                },
                "generated_at": {
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/sqllog.ReportSummary"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "sqllog.ReportSummary": {
            "type": "object",
            "properties": {
                "anomaly_count": {
                    "type": "integer"
                },
                "by_db": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "from": {
                    "type": "string"
                },
                "suggestion_count": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                },
                "total_queries": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}